sort parsing files
