+++ maps/standard/all_shapes.rt
@@ -3,12 +3,7 @@ A 0				255,255,255
 
 c 0,2.5,-10		0,0,1			65
 
-l 0,30,0					1	150,220,190
-l 20,8,-20 					1	190,255,160
-l -20,8,-20 				1 	255,160,190
 pl 0,0,0 		0,1,0 			128,240,128
 pl 0,0,5 		0,0,1 			190,190,190
-sq -1.5,2.5,0 		1,1,0.4 	3	240,128,128
 sp 1.5,2.5,0 					3 	128,128,240
 cy -4.5,2.5,0 	-0.2,-0.3,-1 	3 5 190,255,160
-tr 4,1,1	4.5,5,-1	5,2,0	255,255,255
