+++ maps/standard/mandatory.rt
@@ -1,4 +1,4 @@
-R 1280 720
+R 9000 720
 
 A 0.2 255,255,255
 c -50,1,230 0.,0.,0. 70
